+++ src/Enums/DatetimeZoneAbbreviationEnum.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Shergela\Validations\Enums;
+
+enum DatetimeZoneAbbreviationEnum: string
+{
+    public const AFRICA = 'Africa';
+    public const AMERICA = 'America';
+    public const ANTARCTICA = 'Antarctica';
+    public const ARCTIC = 'Arctic';
+    public const ASIA = 'Asia';
+    public const ATLANTIC = 'Atlantic';
+    public const AUSTRALIA = 'Australia';
+    public const EUROPE = 'Europe';
+    public const INDIAN = 'Indian';
+    public const PACIFIC = 'Pacific';
+}