+++ src/Enums/ValidationDateEnum.php
@@ -18,4 +18,5 @@ enum ValidationDateEnum: string
 
     public const DATE_FORMAT = 'date_format:';
     public const TIMEZONE = 'timezone:';
+    public const TIMEZONE_ALL = 'timezone:all';
 }