+++ src/Enums/ValidationRegexEnum.php
@@ -28,4 +28,19 @@ enum ValidationRegexEnum: string
      * Numbers
      */
     public const NUMBERS = 'regex:/\pN/u';
+
+    /**
+     * Separate integers by comma (1,2,3,4,5).
+     */
+    public const SEPARATE_INTEGERS_BY_COMMA = "/^\d+(?:,\d+)*$/";
+
+    /**
+     * Separate strings by comma. (ana,gustav,john).
+     */
+    public const SEPARATE_STRINGS_BY_COMMA = "/^[a-z]+(?:,[a-z]+)*$/";
+
+    /**
+     * Separate strings by underscore (ana_gustav_john).
+     */
+    public const SEPARATE_STRINGS_BY_UNDERSCORE = "/^[a-z]+(?:_[a-z]+)*$/";
 }